requirements.txt
Django==1.7
Markdown==2.3.1
Pillow==2.5.3
backports.ssl-match-hostname==3.4.0.2
django-admin-tools==0.5.1
django-allauth==0.18.0
django-appconf==0.6
django-bootstrap3==4.11.0
django-braces==1.4.0
django-classy-tags==0.4
django-compressor==1.3
django-cors-headers==0.12
django-crispy-forms==1.4.0
django-debug-toolbar==1.0.1
django-extensions==1.2.2
django-filebrowser==3.5.7
django-filer==0.9.5
django-filter==0.7
django-import-export==0.1.4
django-jsonify==0.2.1
django-localflavor (1.1)
django-modeltranslation==0.7.3
django-mptt==0.6.1
django-nose==1.2
django-oauth-toolkit==0.7.0
django-polymorphic==0.5.3
django-registration==1.0
django-reversion==1.8.4
django-rosetta==0.7.2
django-sekizai==0.7
django-simple-history==1.3.0
django-smuggler==0.6.0
django-suit==0.2.5
django-tastypie==0.10.0
django-user-accounts==1.0b14
djangorestframework==2.3.9
docutils==0.11
easy-thumbnails==1.4
future==0.9.0
html5lib==1.0b3
mimeparse==0.1.3
nose==1.3.0
oauthlib==0.6.1
pisa==3.0.33
pycrypto==2.6
pyparsing==2.0.1
python-dateutil==2.1
python-openid==2.2.5
pytz==2013.7
requests==2.2.1
requests-oauthlib==0.4.0
six==1.4.1
sorl-thumbnail==11.12
sqlparse==0.1.11
tablib==0.9.11
tornado==4.0.2
tzlocal==1.0
urwid==1.3.0
wheel==0.24.0
